Move overlay value of slider in to slider component
