+++ UnstoppableWallet/UnstoppableWallet/Modules/Send/SubModules/SendFeePriority/SendFeePriorityView.swift
@@ -1,7 +1,6 @@
 import UIKit
 import SnapKit
 import UIExtensions
-import HUD
 
 class SendFeePriorityView: UIView {
     let delegate: ISendFeePriorityViewDelegate
@@ -11,7 +10,6 @@ class SendFeePriorityView: UIView {
 
     private let feeSliderWrapper = FeeSliderWrapper()
     private let selectableValueView = SelectableValueView(title: "send.tx_speed".localized)
-    private let feeRateView = FeeSliderValueView()
 
     private let customPriorityUnit: CustomPriorityUnit?
 
@@ -47,9 +45,6 @@ class SendFeePriorityView: UIView {
             maker.top.equalToSuperview()
             maker.leading.trailing.equalToSuperview().inset(CGFloat.margin4x)
         }
-        feeSliderWrapper.onTracking = { [weak self] value, position in
-            self?.onTracking(value, position: position)
-        }
         feeSliderWrapper.finishTracking = { [weak self] value in
             self?.finishTracking(value: value)
         }
@@ -64,34 +59,10 @@ class SendFeePriorityView: UIView {
         selectableValueView.delegate = self
         selectableValueView.set(value: delegate.feeRatePriority.title)
 
-        feeRateView.set(descriptionText: customPriorityUnit?.presentationName)
     }
 
     required init?(coder aDecoder: NSCoder) {
         fatalError("not implemented")
-    }
-
-    private func hudConfig(position: CGPoint) -> HUDConfig {
-        var feeConfig = HUDConfig()
-
-        feeConfig.appearStyle = .alphaAppear
-        feeConfig.style = .banner(.top)
-        feeConfig.absoluteInsetsValue = true
-        feeConfig.userInteractionEnabled = true
-        feeConfig.hapticType = .none
-        feeConfig.blurEffectStyle = nil
-        feeConfig.blurEffectIntensity = nil
-        feeConfig.borderColor = .themeSteel20
-        feeConfig.borderWidth = .heightOnePixel
-        feeConfig.exactSize = true
-        feeConfig.preferredSize = CGSize(width: 74, height: 48)
-        feeConfig.cornerRadius = CGFloat.cornerRadius2x
-        feeConfig.handleKeyboard = .none
-        feeConfig.inAnimationDuration = 0
-        feeConfig.outAnimationDuration = 0
-
-        feeConfig.hudInset = convert(CGPoint(x: position.x - center.x, y: -feeConfig.preferredSize.height - CGFloat.margin2x), to: nil)
-        return feeConfig
     }
 
     private func convert(_ value: Int, toPresent: Bool = true) -> Int {
@@ -105,16 +76,7 @@ class SendFeePriorityView: UIView {
         convert(range.lowerBound, toPresent: toPresent)...convert(range.upperBound, toPresent: toPresent)
     }
 
-    private func onTracking(_ value: Int, position: CGPoint) {
-        HUD.instance.config = hudConfig(position: position)
-
-        feeRateView.set(value: "\(value)")
-        HUD.instance.showHUD(feeRateView)
-    }
-
     private func finishTracking(value: Int) {
-        HUD.instance.hide()
-
         let realRange = convert(feeSliderWrapper.sliderRange, toPresent: false)
         delegate.selectCustom(feeRatePriority: .custom(value: convert(value, toPresent: false), range: realRange))
     }
@@ -141,7 +103,7 @@ extension SendFeePriorityView: ISendFeePriorityView {
 
     func set(customFeeRateValue: Int, customFeeRateRange: ClosedRange<Int>) {
         let presentationRange = convert(customFeeRateRange)
-        feeSliderWrapper.set(value: convert(customFeeRateValue), range: presentationRange)
+        feeSliderWrapper.set(value: convert(customFeeRateValue), range: presentationRange, description: customPriorityUnit?.presentationName)
     }
 
     func set(duration: TimeInterval?) {
